Abstracting useOnSwipe hook
